requirements.txt
pymupdf
chromadb
sentence-transformers
groq
tavily-python
python-dotenv
flask
flask-cors
huggingface_hub
requests
ddgs